+++ tools/vscode-extension/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to the "bamboo" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [0.0.2] - 2019-10-29
+### Changed
+- Renamed instances of `bamboo` to `cadence`
+- Configuration option to start language server now supports arguments
+
 ## [Unreleased]
 
 - Initial release